Adding basic repository and revision code.
This code will load, but is mostly un-tested. Roots represent filesystem root
objects, which are Revisions and Transactions (later). Revisions inherit from
Root and provide access to files, folders, and properties at some immutable
revision within a SVN Repo. Transaction objects should (eventually) behave the
same way and will encapuslate both the svn_fs_txn_t and a svn_fs_root_t.

The Repo class encapsulates both svn_repos_t and its associated svn_fs_t. In
this high-level library, there should be no need for the distinction. Right
now, a new sub-pool is allocated when a Repo is created or opened. This pool
should have the same lifetime as the Repo (and FileSystem), which will be freed
when its pool is destroyed.