Amending Utils#content_for.
In the previous commit, content_for was a mix between the two current methods,
content_for and content_from. Content_from is used when an out pointer is used
and it will read a pointer or integer out of the pointer that it is given. This
is what :returning clauses in bind calls use. Content_for is simply the reverse
of pointer_for. When objects are stored using pointer_for, content_for should
be used to read those objects.

Also included in this commit is a partial implementation of AprArrays.