Fixing Repo#history because structs are reused.
The previous version returned 4 copies of the last log entry object because all
of the pointers passed to the log entry receiver function pointed to the same
object (either because the struct was always re-used or because the memory was
freed after every call). Since it was clear that the values could not be simply
collected, the solution was to create a multi-level hash out of the log entry
data, which is implemented in the Log::EntryStruct#to_h and similar methods. It
was not possible to simply copy the Log::EntryStruct into ruby-managed memory
because all sub-structures would have to be individually copied as well.

Repository history is now working as expected.