Updating errors for better classes, messages.
Error classes will now be dynamically generated from SVN error "best" messages,
and error messages will be the "root cause" message. This produces what appear
to be better errors and specific classes that can be caught... the classes may
not exist until they are thrown, which is still a bug. With the new structure,
"Svn::RepositoryCreationFailedError: '/tmp' exists and is non-empty" is raised
instead of "Svn::Error: Repository creation failed".

This also moves Svn::CError over to using bind for instance methods, which
required adding Extensions to FFI::Structs in Utils.