Adding better error handling.
Turning the "best" error message into an error class was a bad idea. This will
now turn generic error messages for the error code into dynamically created
Svn::Error sub-classes, to cover any errors that may be returned, but known
errors will be added statically using the new Svn::Error.add method. This way,
the known errors exist in the library for rescue statements and there is an
easy way to list the errors and control the error naming:
 Svn::Error.add(2, PathNotFoundError)