Adding Svn::Repo.open spec and updating errors.
The spec is now up to date with the previous commit's error changes;
specifically, using PathNotFoundError and DirectoryNotEmptyError. Since the
RepositoryCreationFailedError no longer exists, changed the expected error for
nil paths to be an ArgumentError.

Added spec code for Svn::Repo.open, which mainly complains about invalid paths,
and added a "not nil" check for repository objects in both .open and .create
specs.